Add additive migration for text_settings reconciliation

+++ migrations/versions/b3c4d5e6f7a8_reconcile_text_settings_table.py
@@ -1,8 +1,8 @@
-"""add text settings table
+"""reconcile text settings table
 
-Revision ID: a8c9d0e1f2a3
-Revises: 3979d4be8acf
-Create Date: 2026-04-12 00:00:00.000000
+Revision ID: b3c4d5e6f7a8
+Revises: c4d5e6f7a8b9
+Create Date: 2026-04-18 00:00:00.000000
 """
 
 from alembic import op
@@ -10,13 +10,19 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision = 'a8c9d0e1f2a3'
-down_revision = '3979d4be8acf'
+revision = 'b3c4d5e6f7a8'
+down_revision = 'c4d5e6f7a8b9'
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
+    bind = op.get_bind()
+    inspector = sa.inspect(bind)
+
+    if inspector.has_table('text_settings'):
+        return
+
     op.create_table(
         'text_settings',
         sa.Column('id', sa.Integer(), nullable=False),
@@ -28,4 +34,8 @@ def upgrade():
 
 
 def downgrade():
-    op.drop_table('text_settings')
+    bind = op.get_bind()
+    inspector = sa.inspect(bind)
+
+    if inspector.has_table('text_settings'):
+        op.drop_table('text_settings')
